try to make color transfer

+++ libs/gpuimage/src/main/cpp/yuv-decoder.cpp
@@ -51,23 +51,23 @@ uint32_t bgra_to_argb(uint32_t bgra) {
     return argb;
 }
 
-double pivotXyzComponent(double component) {
+double pivotXyzComponent(float component) {
     return component > XYZ_EPSILON ? pow(component, 1.0 / 3.0) : (XYZ_KAPPA * component + 16) /
                                                                  116.0;
 }
 
-void colorToLAB(double red, double green, double blue, double lab[3]) {
-    double r = red;
-    double g = green;
-    double b = blue;
+void colorToLAB(float red, float green, float blue, float lab[3]) {
+    float r = red;
+    float g = green;
+    float b = blue;
 
     r = (r > 0.04045) ? pow((r + 0.055) / 1.055, 2.4) : r / 12.92;
     g = (g > 0.04045) ? pow((g + 0.055) / 1.055, 2.4) : g / 12.92;
     b = (b > 0.04045) ? pow((b + 0.055) / 1.055, 2.4) : b / 12.92;
 
-    double x = r * 0.4124 + g * 0.3576 + b * 0.1805;
-    double y = r * 0.2126 + g * 0.7152 + b * 0.0722;
-    double z = r * 0.0193 + g * 0.1192 + b * 0.9505;
+    float x = r * 0.4124 + g * 0.3576 + b * 0.1805;
+    float y = r * 0.2126 + g * 0.7152 + b * 0.0722;
+    float z = r * 0.0193 + g * 0.1192 + b * 0.9505;
 
     x = x / XYZ_WHITE_REFERENCE_X;
     y = y / XYZ_WHITE_REFERENCE_Y;
@@ -83,14 +83,14 @@ void colorToLAB(double red, double green, double blue, double lab[3]) {
     lab[2] = 200.0 * (y - z);
 }
 
-void labToColor(double l, double a, double b, double rgb[3]) {
-    double y = (l + 16.0) / 116.0;
-    double x = a / 500.0 + y;
-    double z = y - b / 200.0;
+void labToColor(float l, float a, float b, float rgb[3]) {
+    float y = (l + 16.0) / 116.0;
+    float x = a / 500.0 + y;
+    float z = y - b / 200.0;
 
-    double x3 = pow(x, 3.0);
-    double y3 = pow(y, 3.0);
-    double z3 = pow(z, 3.0);
+    float x3 = pow(x, 3.0);
+    float y3 = pow(y, 3.0);
+    float z3 = pow(z, 3.0);
 
     x = (x3 > XYZ_EPSILON) ? x3 : (116.0 * x - 16.0) / XYZ_KAPPA;
     y = (y3 > XYZ_EPSILON) ? y3 : (116.0 * y - 16.0) / XYZ_KAPPA;
@@ -101,26 +101,26 @@ void labToColor(double l, double a, double b, double rgb[3]) {
     z *= XYZ_WHITE_REFERENCE_Z;
 
     // Преобразование XYZ в RGB
-    double r = x * 3.2406 + y * -1.5372 + z * -0.4986;
-    double g = x * -0.9689 + y * 1.8758 + z * 0.0415;
-    double bl = x * 0.0557 + y * -0.2040 + z * 1.0570;
+    float r = x * 3.2406 + y * -1.5372 + z * -0.4986;
+    float g = x * -0.9689 + y * 1.8758 + z * 0.0415;
+    float bl = x * 0.0557 + y * -0.2040 + z * 1.0570;
 
     r = (r > 0.0031308) ? 1.055 * pow(r, 1.0 / 2.4) - 0.055 : 12.92 * r;
     g = (g > 0.0031308) ? 1.055 * pow(g, 1.0 / 2.4) - 0.055 : 12.92 * g;
     bl = (bl > 0.0031308) ? 1.055 * pow(bl, 1.0 / 2.4) - 0.055 : 12.92 * bl;
 
     // Ограничение значений в диапазоне [0, 1]
-    rgb[0] = std::min(std::max(r, 0.0), 1.0);
-    rgb[1] = std::min(std::max(g, 0.0), 1.0);
-    rgb[2] = std::min(std::max(bl, 0.0), 1.0);
+    rgb[0] = std::min(std::max(r, 0.0f), 1.0f);
+    rgb[1] = std::min(std::max(g, 0.0f), 1.0f);
+    rgb[2] = std::min(std::max(bl, 0.0f), 1.0f);
 }
 
 void
-calculateMeanAndStdLAB(JNIEnv *env, void *pixels, int width, int height, int stride, double mean[3],
-                       double std[3]) {
+calculateMeanAndStdLAB(JNIEnv *env, void *pixels, int width, int height, int stride, float mean[3],
+                       float std[3]) {
     int pixelCount = stride * height;
-    double sumSq[3] = {0.0, 0.0, 0.0};
-    double sum[3] = {0.0, 0.0, 0.0};
+    float sumSq[3] = {0.0f, 0.0f, 0.0f};
+    float sum[3] = {0.0f, 0.0f, 0.0f};
 
     for (int y = 0; y < height; ++y) {
         auto pixelsComp = reinterpret_cast<uint8_t *>(reinterpret_cast<uint8_t *>(pixels) +
@@ -128,11 +128,11 @@ calculateMeanAndStdLAB(JNIEnv *env, void *pixels, int width, int height, int str
         int x = 0;
 
         for (; x < width; ++x) {
-            double r = pixelsComp[0] / 255.0;
-            double g = pixelsComp[1] / 255.0;
-            double b = pixelsComp[2] / 255.0;
+            float r = pixelsComp[0] / 255.0;
+            float g = pixelsComp[1] / 255.0;
+            float b = pixelsComp[2] / 255.0;
 
-            double lab[3];
+            float lab[3];
             colorToLAB(r, g, b, lab);
             sum[0] += lab[0];
             sum[1] += lab[1];
@@ -152,11 +152,11 @@ calculateMeanAndStdLAB(JNIEnv *env, void *pixels, int width, int height, int str
         int x = 0;
 
         for (; x < width; ++x) {
-            double r = pixelsComp[0] / 255.0;
-            double g = pixelsComp[1] / 255.0;
-            double b = pixelsComp[2] / 255.0;
+            float r = pixelsComp[0] / 255.0;
+            float g = pixelsComp[1] / 255.0;
+            float b = pixelsComp[2] / 255.0;
 
-            double lab[3];
+            float lab[3];
             colorToLAB(r, g, b, lab);
 
             sumSq[0] += (lab[0] - mean[0]) * (lab[0] - mean[0]);
@@ -216,14 +216,14 @@ Java_jp_co_cyberagent_android_gpuimage_GPUImageNativeLibrary_transferPalette(
                                                  static_cast<jint>(targetHeight), rgba8888Obj);
 
     // Вычисление среднего и стандартного отклонения для источника в LAB
-    double sourceMean[3] = {0.0, 0.0, 0.0};
-    double sourceStd[3] = {0.0, 0.0, 0.0};
+    float sourceMean[3] = {0.0, 0.0, 0.0};
+    float sourceStd[3] = {0.0, 0.0, 0.0};
     calculateMeanAndStdLAB(env, sourcePixels, sourceWidth, sourceHeight, sourceInfo.stride,
                            sourceMean, sourceStd);
 
     // Вычисление среднего и стандартного отклонения для цели в LAB
-    double targetMean[3] = {0.0, 0.0, 0.0};
-    double targetStd[3] = {0.0, 0.0, 0.0};
+    float targetMean[3] = {0.0, 0.0, 0.0};
+    float targetStd[3] = {0.0, 0.0, 0.0};
     calculateMeanAndStdLAB(env, targetPixels, targetWidth, targetHeight, targetInfo.stride,
                            targetMean, targetStd);
 
@@ -243,25 +243,25 @@ Java_jp_co_cyberagent_android_gpuimage_GPUImageNativeLibrary_transferPalette(
         int x = 0;
 
         for (; x < targetWidth; ++x) {
-            double r = pixelsComp[0] / 255.0;
-            double g = pixelsComp[1] / 255.0;
-            double b = pixelsComp[2] / 255.0;
+            float r = pixelsComp[0] / 255.0;
+            float g = pixelsComp[1] / 255.0;
+            float b = pixelsComp[2] / 255.0;
 
-            double lab[3];
+            float lab[3];
             colorToLAB(r, g, b, lab);
 
-            double newL = ((lab[0] - targetMean[0]) * (sourceStd[0] / targetStd[0]) +
+            float newL = ((lab[0] - targetMean[0]) * (sourceStd[0] / targetStd[0]) +
                            sourceMean[0]);
-            double newA = ((lab[1] - targetMean[1]) * (sourceStd[1] / targetStd[1]) +
+            float newA = ((lab[1] - targetMean[1]) * (sourceStd[1] / targetStd[1]) +
                            sourceMean[1]);
-            double newB = ((lab[2] - targetMean[2]) * (sourceStd[2] / targetStd[2]) +
+            float newB = ((lab[2] - targetMean[2]) * (sourceStd[2] / targetStd[2]) +
                            sourceMean[2]);
 
-            double finalL = lab[0] + intensity * (newL - lab[0]);
-            double finalA = lab[1] + intensity * (newA - lab[1]);
-            double finalB = lab[2] + intensity * (newB - lab[2]);
+            float finalL = lab[0] + intensity * (newL - lab[0]);
+            float finalA = lab[1] + intensity * (newA - lab[1]);
+            float finalB = lab[2] + intensity * (newB - lab[2]);
 
-            double rgb[3];
+            float rgb[3];
             labToColor(finalL, finalA, finalB, rgb);
 
             float red = std::clamp((float) rgb[0] * 255.0f, 0.0f, 255.0f);
